third_party/blink/web_tests/external/wpt/custom-elements/registries/element-mutation-expected.txt

This is a testharness.js-based test.
[FAIL] An element with global registry should not change its registry when moved into a shadow tree with scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when moved out of the shadow tree.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] An element with scoped registry should not change its registry when moved into another shadow tree with different scoped registry.
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] Declarative shadow DOM with shadowrootcustomelementregistry attribute without registry initialized should remain null registry after adoption.
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
